Move to separate test class all host network mode related tests
